// apps/sync-service/src/openapi/order-status-confirm.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * SMILE
 * OpenAPI spec version: 1.0.0
 */
import axios from "axios"
import type { AxiosRequestConfig, AxiosResponse } from "axios"

export type PutV2OrderConfirmBodyOrderItemsItem = {
  id?: number
  order_id?: number
  material_id?: number
  qty?: number
  confirmed_qty?: number
}

export type PutV2OrderConfirmBody = {
  order_items: PutV2OrderConfirmBodyOrderItemsItem[]
  comment?: string | null
}

export type PutV2OrderConfirm200 = { [key: string]: unknown }

export const getSmile = () => {
  /**
   * @summary Konfirmasi Order
   */
  const putV2OrderConfirm = <TData = AxiosResponse<PutV2OrderConfirm200>>(
    orderId: number,
    putV2OrderConfirmBody: PutV2OrderConfirmBody,
    options?: AxiosRequestConfig
  ): Promise<TData> => {
    return axios.put(
      `/v2/order/${orderId}/confirm`,
      putV2OrderConfirmBody,
      options
    )
  }

  return { putV2OrderConfirm }
}
export type PutV20OrderConfrirmResult = AxiosResponse<PutV2OrderConfirm200>